+++ src/main/java/com/meoguri/linkocean/controller/category/CategoryController.java
@@ -1,10 +1,10 @@
-package com.meoguri.linkocean.controller;
+package com.meoguri.linkocean.controller.category;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.meoguri.linkocean.controller.dto.category.AllCategoriesResponse;
+import com.meoguri.linkocean.controller.category.dto.AllCategoriesResponse;
 import com.meoguri.linkocean.domain.bookmark.service.CategoryService;
 
 import lombok.RequiredArgsConstructor;
