+++ src/main/java/com/meoguri/linkocean/controller/category/dto/AllCategoriesResponse.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.controller.dto.category;
+package com.meoguri.linkocean.controller.category.dto;
 
 import java.util.List;
